fix: refactor inject and table building functions
