+++ src/lib/datetime.ts
@@ -24,15 +24,15 @@ export function unixTimestampToLocale (unixTimestamp: number): string {
   const date = new Date(unixTimestamp * 1000)
   const locale = navigator.language || 'sv-SE'
 
-  const options: Intl.DateTimeFormatOptions = {
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric',
-      hour: 'numeric',
-      minute: 'numeric',
-      second: 'numeric',
-      timeZoneName: 'short'
-  }
+  const options = {
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric',
+    hour: 'numeric',
+    minute: 'numeric',
+    second: 'numeric',
+    timeZoneName: 'short',
+  } as Intl.DateTimeFormatOptions
 
   return new Intl.DateTimeFormat(locale, options).format(date)
 }
