+++ src/renderers/stats.ts
@@ -1,4 +1,5 @@
 import { fetchAchievements, fetchOrders, fetchSupplyDrops } from '../lib/api'
+import { addDataToDiv, findInjectPath } from '../lib/builders'
 import { getCachedUser } from '../lib/userIdCache'
 import { findCategoriesByPeriod } from '../reducers/categories'
 import { type ExperienceStats, getExperienceStats } from '../reducers/experience'
@@ -6,54 +7,6 @@ import { type HoarderEntry, findTopHoarderCheevoStats } from '../reducers/hoarde
 import { findOrdersPerMonth } from '../reducers/orders'
 import { getStoreStats, type StoreSum } from '../reducers/stores'
 import { type Streaks, findStreaks } from '../reducers/streaks'
-
-function addDataToDiv (headerText: string, domObject: Element): HTMLDivElement {
-  const div = document.createElement('div')
-  div.className = 'order my-4'
-
-  const table = document.createElement('table')
-  table.className = 'table table-condensed'
-
-  const tbody = document.createElement('tbody')
-
-  const tr = document.createElement('tr')
-  tr.className = 'order-id-wrap'
-
-  const td = document.createElement('td')
-  td.textContent = headerText
-
-  tr.appendChild(td)
-  tbody.appendChild(tr)
-  table.appendChild(tbody)
-  div.appendChild(table)
-
-  const div1 = document.createElement('div')
-  const div2 = document.createElement('div')
-  const orderProgression = document.createElement('div')
-  const innerContainer = document.createElement('div')
-  const orderStatusEvent = document.createElement('div')
-  const icon = document.createElement('div')
-  const header = document.createElement('h3')
-  const secondary = document.createElement('div')
-
-  div1.appendChild(div2)
-  div2.appendChild(orderProgression)
-  orderProgression.appendChild(innerContainer)
-  innerContainer.appendChild(orderStatusEvent)
-  orderStatusEvent.appendChild(icon)
-  orderStatusEvent.appendChild(header)
-  orderStatusEvent.appendChild(secondary)
-  secondary.appendChild(domObject)
-
-  header.className = 'level-two-heading'
-  icon.className = 'icon'
-
-  header.textContent = ''
-
-  div.appendChild(div1)
-
-  return div
-}
 
 function addSortingFunctionality (table: HTMLTableElement, headers: string[]): void {
   const thead = table.querySelector('thead') as HTMLTableSectionElement
@@ -360,17 +313,6 @@ function generateExperienceTable (jsonData: ExperienceStats): HTMLTableElement {
   table.appendChild(tbody)
 
   return table
-}
-function findInjectPath (paths: string[]): HTMLElement | null {
-  let dom = null
-  paths.forEach(path => {
-    const d = document.querySelector(path)
-    if (d) {
-      dom = d
-    }
-  })
-
-  return dom
 }
 
 function generateHoarderTable (jsonData: HoarderEntry[]): HTMLTableElement {
